# swagger/swagger.yaml
consumes:
- application/json
- application/xml
- application/gob
- application/x-gob
definitions:
  CreateUserPayload:
    example:
      name: John
    properties:
      name:
        description: name
        example: John
        type: string
    required:
    - name
    title: CreateUserPayload
    type: object
  UpdateUserPayload:
    example:
      name: John
    properties:
      name:
        description: name
        example: John
        type: string
    required:
    - name
    title: UpdateUserPayload
    type: object
  User:
    description: user (default view)
    example:
      id: 4.9096287e+15
      id_str: "4909628655665152"
      name: John
    properties:
      id:
        description: id(int64)
        example: 4.9096287e+15
      id_str:
        description: id(string)
        example: "4909628655665152"
        type: string
      name:
        description: name
        example: John
        type: string
    required:
    - id
    - id_str
    - name
    title: 'Mediatype identifier: application/vnd.user+json; view=default'
    type: object
  UserCollection:
    description: UserCollection is the media type for an array of User (default view)
    example:
    - id: 4.9096287e+15
      id_str: "4909628655665152"
      name: John
    - id: 4.9096287e+15
      id_str: "4909628655665152"
      name: John
    items:
      $ref: '#/definitions/User'
    title: 'Mediatype identifier: application/vnd.user+json; type=collection; view=default'
    type: array
  error:
    description: Error response media type (default view)
    example:
      code: invalid_value
      detail: Value of ID must be an integer
      id: 3F1FKVRR
      meta:
        timestamp: 1.458609e+09
      status: "400"
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: invalid_value
        type: string
      detail:
        description: a human-readable explanation specific to this occurrence of the
          problem.
        example: Value of ID must be an integer
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: 3F1FKVRR
        type: string
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error.
        example:
          timestamp: 1.458609e+09
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: "400"
        type: string
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
host: localhost:8080
info:
  description: A simple appengine example
  title: The appengine example
  version: ""
paths:
  /swagger.json:
    get:
      operationId: swagger#/swagger.json
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
      schemes:
      - http
      summary: Download ../swagger/swagger.json
  /swagger/{filepath}:
    get:
      operationId: swagger#/swagger/*filepath
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
        "404":
          description: File not found
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: Download ../static/swagger/
  /users:
    get:
      description: list
      operationId: User#list
      parameters:
      - description: name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/vnd.goa.error
      - application/vnd.goa.error
      - application/vnd.goa.error
      - application/vnd.user+json; type=collection
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserCollection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: list User
      tags:
      - User
    post:
      description: create
      operationId: User#create
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/CreateUserPayload'
      produces:
      - application/vnd.goa.error
      - application/vnd.user+json
      - application/vnd.goa.error
      - application/vnd.goa.error
      - application/vnd.goa.error
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: create User
      tags:
      - User
  /users/{id}:
    delete:
      description: delete
      operationId: User#delete
      parameters:
      - description: id(int64)
        in: path
        name: id
        required: true
        type: any
      produces:
      - application/vnd.goa.error
      - application/vnd.goa.error
      - application/vnd.user+json
      - application/vnd.goa.error
      - application/vnd.goa.error
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: delete User
      tags:
      - User
    get:
      description: show
      operationId: User#show
      parameters:
      - description: id(int64)
        in: path
        name: id
        required: true
        type: any
      produces:
      - application/vnd.goa.error
      - application/vnd.goa.error
      - application/vnd.goa.error
      - application/vnd.user+json
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: show User
      tags:
      - User
    put:
      description: update
      operationId: User#update
      parameters:
      - description: id(int64)
        in: path
        name: id
        required: true
        type: any
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/UpdateUserPayload'
      produces:
      - application/vnd.goa.error
      - application/vnd.goa.error
      - application/vnd.goa.error
      - application/vnd.user+json
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: update User
      tags:
      - User
produces:
- application/json
- application/xml
- application/gob
- application/x-gob
schemes:
- http
swagger: "2.0"